Refactor TlsChannelStream
Pull up methods to a new abstract base class that doesn't depend on
the TlsChannel implementation.  This base class can be extended by
AsynchronousSocketChannelStream as well.

JAVA-3038
